+++ containers/payments/PaymentVerificationModal.js
@@ -1,43 +1,39 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { FormModal, Alert, Loader, useNotifications } from 'react-components';
+import { FormModal, Alert, Loader, useNotifications, useLoading, useApi } from 'react-components';
 import { c } from 'ttag';
 import tabSvg from 'design-system/assets/img/pm-images/tab.svg';
 
-import { toParams } from './paymentTokenHelper';
-import PaymentVerificationButton from './PaymentVerificationButton';
+import { toParams, process } from './paymentTokenHelper';
 
 const PaymentVerificationModal = ({ params, token, approvalURL, onSubmit, ...rest }) => {
+    const [loading, withLoading] = useLoading();
+    const api = useApi();
     const { createNotification } = useNotifications();
-    const [loading, setLoading] = useState(false);
     const title = loading ? c('Title').t`Payment verification in progress` : c('Title').t`Payment verification`;
 
     const handleSubmit = async () => {
-        onSubmit(toParams(params, token));
-        rest.onClose();
-    };
+        const tab = window.open(approvalURL);
 
-    const handleError = (error) => {
-        rest.onClose();
+        try {
+            await process({ Token: token, api, tab });
+            onSubmit(toParams(params, token));
+            rest.onClose();
+        } catch (error) {
+            rest.onClose();
 
-        // if not coming from API error
-        if (error.message && !error.config) {
-            createNotification({ text: error.message, type: 'error' });
+            // if not coming from API error
+            if (error.message && !error.config) {
+                createNotification({ text: error.message, type: 'error' });
+            }
         }
     };
 
     return (
         <FormModal
             title={title}
-            submit={
-                <PaymentVerificationButton
-                    approvalURL={approvalURL}
-                    token={token}
-                    onError={handleError}
-                    onSubmit={handleSubmit}
-                    onLoading={setLoading}
-                >{c('Action').t`Verify payment`}</PaymentVerificationButton>
-            }
+            submit={c('Action').t`Verify payment`}
+            onSubmit={() => withLoading(handleSubmit())}
             small={true}
             {...rest}
         >
